+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Bug Fixes
 
+- Don't null out the statusbar item spacings which caused the statusbar items to have 0 padding and look weird.
 - Fix `ColorDialog` not showing up the colored buttons.
 - Make `RecentFilesPicker` be flat structured by default instead of hierarchical (it makes more sense to have flat structure)
 - Handle `Esc` key to quit open pickers.